lpc: Handle errors gracefully
The lpc decoder blindly transitions to GET ADDR state even if it
encounters an error or a DMA packet (which doesn't have a GET ADDR
state). In handle_get_addr() we then set addr_nibbles=0, do a negative
shift and die.

Fix this in two ways - firstly catch these issues in handle_get_ct_dr()
and do not advance the state machine on. Secondly just assert if we
somehow ended up in GET ADDR state for a DMA transaction, which should
never happen now.